added login feature and moved the add quote button to the center
